+++ src/wondergui/widgets/overlays/wg_tooltipoverlay.h
@@ -63,7 +63,7 @@ namespace wg
 			int				id = 0;
 			MarkPolicy		markPolicy = MarkPolicy::AlphaTest;
 			bool			pickable = false;
-			int				pickCategory = 0;
+			uint8_t			pickCategory = 0;
 			PointerStyle	pointer = PointerStyle::Undefined;
 			Position		position;
 			bool			selectable = true;
