+++ src/wondergui/widgets/wg_areachart.h
@@ -187,7 +187,7 @@ namespace wg
 
 			MarkPolicy		markPolicy = MarkPolicy::AlphaTest;
 			bool			pickable = false;
-			int				pickCategory = 0;
+			uint8_t			pickCategory = 0;
 			PointerStyle	pointer = PointerStyle::Undefined;
 
 			Placement		rightLabelPlacement = Placement::East;
