+++ src/wondergui/widgets/wg_popupopener.h
@@ -56,7 +56,7 @@ namespace wg
 			MouseButton		mouseButton		= MouseButton::Left;
 			bool			openOnHover		= false;
 			bool			pickable		= false;
-			int				pickCategory	= 0;
+			uint8_t			pickCategory	= 0;
 			PointerStyle	pointer			= PointerStyle::Undefined;
 			Widget_p		popup;
 			Border			popupOverflow;
